backend/requirements.txt
python-telegram-bot==21.5
paho-mqtt==1.6.1
geopy==2.4.0
